Cleanup code + reimplement address forget
